Swap wordcounts for email signup
